requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements/base.in
#
asgiref==3.6.0
    # via django
django==4.1.6
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
    #   django-configurations
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
django-configurations==2.4
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
django-cors-headers==3.7.0
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
django-extensions==3.2.1
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
django-filter==23.2
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
djangorestframework==3.14.0
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
pillow==9.4.0
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
psycopg2==2.9.5
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
pyjwt==2.6.0
    # via djangorestframework-simplejwt
pytz==2023.3
    # via djangorestframework
sqlparse==0.4.3
    # via django
